Add --resolve to target-query (#485)
* Changes the target.resolve() functionality to return a TargetPath
* Create modifiers that can change records before printing them.

(DIS-2624)
